refactor: organize test files
